break metadata in different component
