+++ console/console-init/ui/src/modules/iot-device/components/ReviewDevice/ReviewDevice.tsx
@@ -66,12 +66,7 @@ const ReviewDevice: React.FunctionComponent<IReviewDeviceProps> = ({
               MetaData
             </Title>
             <br />
-            <ReviewMetaData
-              defaults={defaults}
-              ext={ext}
-              label={"Default properties parameter"}
-            />
-            <br />
+            <ReviewMetaData defaults={defaults} ext={ext} />
             <br />
             <br />
           </>
