+++ console/console-init/ui/src/modules/iot-device/components/ReviewDevice/ReviewMetaData.tsx
@@ -1,111 +1,24 @@
-import React, { useState } from "react";
-import { Grid, GridItem, TextInput, Button } from "@patternfly/react-core";
+/*
+ * Copyright 2020, EnMasse authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+
+import React from "react";
+import { Grid, GridItem } from "@patternfly/react-core";
 import { convertJsonToMetadataOptions } from "utils";
 import { StyleSheet, css } from "aphrodite";
-import { AngleRightIcon, AngleDownIcon } from "@patternfly/react-icons";
+import { MetaDataReviewRows } from "./MetaDataReviewRows";
 
 const styles = StyleSheet.create({
-  grid_align: { marginTop: 10, marginRight: 5, marginLeft: 5 },
-  grid_button_align: {
-    textAlign: "right",
-    marginTop: 10,
-    marginRight: 5,
-    marginLeft: 5
-  },
   text_center_align: { textAlign: "center" }
 });
-interface IMetaDataReviewRowProps {
-  values: any[];
-  prevkey?: string;
-}
-
-const MetaDataReviewRow: React.FC<IMetaDataReviewRowProps> = ({
-  values,
-  prevkey
-}) => {
-  return (
-    <>
-      {values.map((val, index) => (
-        <GetDevice value={val} prevKey={prevkey} index={index} />
-      ))}
-    </>
-  );
-};
-
-interface IRowProps {
-  value: any;
-  prevKey?: string;
-  index?: number;
-}
-
-const GetDevice: React.FunctionComponent<IRowProps> = ({
-  value,
-  prevKey,
-  index
-}) => {
-  const [isVisible, setIsVisible] = useState<boolean>(false);
-
-  const onToggle = () => {
-    setIsVisible(!isVisible);
-  };
-  const renderGridItem = (value: string) => {
-    return (
-      <GridItem span={3} className={css(styles.grid_align)}>
-        <TextInput
-          key={value + index}
-          id={value + index}
-          type="text"
-          value={value}
-          isReadOnly={true}
-        />
-      </GridItem>
-    );
-  };
-
-  const hasObjectValue: boolean =
-    value.type === "array" || value.type === "object";
-  const key: string = prevKey
-    ? value.key !== ""
-      ? prevKey + "/" + value.key
-      : prevKey
-    : value.key;
-  return (
-    <>
-      <Grid>
-        <GridItem span={1} className={css(styles.grid_button_align)}>
-          {hasObjectValue && (
-            <Button
-              variant="plain"
-              id={key}
-              style={{ textAlign: "right" }}
-              onClick={onToggle}
-            >
-              {isVisible ? <AngleDownIcon /> : <AngleRightIcon />}
-            </Button>
-          )}
-        </GridItem>
-        {renderGridItem(key)}
-        {renderGridItem(value.typeLabel)}
-        {renderGridItem(hasObjectValue ? "" : value.value)}
-        {hasObjectValue && isVisible && (
-          <>
-            <MetaDataReviewRow values={value.value} prevkey={key} />
-          </>
-        )}
-      </Grid>
-    </>
-  );
-};
-
 interface IReviewMetaDataProrps {
   defaults?: any[];
   ext?: any[];
-  label: string;
 }
 const ReviewMetaData: React.FunctionComponent<IReviewMetaDataProrps> = ({
   defaults,
-  ext,
-  label
+  ext
 }) => {
   if (ext === undefined && defaults === undefined) {
     return <>--</>;
@@ -115,7 +28,7 @@ const ReviewMetaData: React.FunctionComponent<IReviewMetaDataProrps> = ({
   if (convertedDefaultsData.length === 0 && convertedExtData.length === 0) {
     return <>--</>;
   }
-  const renderGrid = (values: any[]) => {
+  const renderGrid = (values: any[], label: string) => {
     return (
       <>
         <Grid>
@@ -130,14 +43,17 @@ const ReviewMetaData: React.FunctionComponent<IReviewMetaDataProrps> = ({
             <b>Value</b>
           </GridItem>
         </Grid>
-        <MetaDataReviewRow values={values} />
+        <MetaDataReviewRows values={values} />
       </>
     );
   };
   return (
     <>
-      {convertedDefaultsData.length > 0 && renderGrid(convertedDefaultsData)}
-      {convertedExtData.length > 0 && renderGrid(convertedExtData)}
+      {convertedDefaultsData.length > 0 &&
+        renderGrid(convertedDefaultsData, "Default properties parameter")}
+      <br />
+      {convertedExtData.length > 0 &&
+        renderGrid(convertedExtData, "Extension parameter")}
     </>
   );
 };
